requirements.txt
httpx==0.27.2
python-telegram-bot==21.6
